# document/spec/openapi.yaml
openapi: 3.1.0
info:
  title: lib4gz LMS Platform Backend API
  description: >
    RESTful API for the lib4gz Learning Management System platform.
    Provides endpoints for user authentication, course management, enrollments,
    modules, lessons, summaries, exercises, questions, and submissions.
  version: 1.0.0
  contact:
    name: lib4gz Team
servers:
  - url: http://localhost:8080/api
    description: Local Development
security:
  - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: access_token
      description: JWT access token
  parameters:
    XUserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: TypeID of the authenticated user, injected by JwtAuthFilter
      schema:
        type: string
        format: typeid
    CourseIdPath:
      name: courseId
      in: path
      required: true
      description: TypeID of the course
      schema:
        type: string
        format: typeid
...
    EnrollmentIdPath:
      name: enrollmentId
      in: path
      required: true
      description: TypeID of the enrollment
      schema:
        type: string
        format: typeid
...
  schemas:
...
    Visibility:
      type: string
      enum:
        - PUBLIC
        - PRIVATE
        - UNLISTED
    EnrollmentRole:
      type: string
      enum:
        - LEARNER
        - TEACHER
    EnrollmentStatus:
      type: string
      enum:
        - PENDING
        - ACTIVE
        - REJECTED
        - INACTIVE
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: typeid
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type:
            - string
            - "null"

    # ---------- Courses ----------
    CreateCourseRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type:
            - string
            - "null"
        visibility:
          $ref: "#/components/schemas/Visibility"
          default: PRIVATE
        settings:
          type: object
          default: {}
    UpdateCourseRequest:
      type: object
      properties:
        title:
          type:
            - string
            - "null"
        description:
          type:
            - string
            - "null"
        visibility:
          oneOf:
            - $ref: "#/components/schemas/Visibility"
            - type: "null"
        settings:
          type:
            - object
            - "null"
    CourseResponse:
      type: object
      properties:
        id:
          type: string
          format: typeid
        code:
          type: string
          description: Unique 10-digit numeric code for course lookup
          pattern: "^\\d{10}$"
        title:
          type: string
        description:
          type:
            - string
            - "null"
        visibility:
          $ref: "#/components/schemas/Visibility"
        createdBy:
          $ref: "#/components/schemas/UserSummary"
        settings:
          type: object
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        moduleCount:
          type:
            - integer
            - "null"
        enrollmentCount:
          type:
            - integer
            - "null"

    # ---------- Enrollments ----------
    EnrollmentRequest:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/EnrollmentRole"
          default: LEARNER
    UpdateEnrollmentRequest:
      type: object
      properties:
        role:
          oneOf:
            - $ref: "#/components/schemas/EnrollmentRole"
            - type: "null"
        status:
          oneOf:
            - $ref: "#/components/schemas/EnrollmentStatus"
            - type: "null"
    EnrollmentResponse:
      type: object
      properties:
        id:
          type: string
          format: typeid
        courseId:
          type: string
          format: typeid
        courseName:
          type: string
        user:
          $ref: "#/components/schemas/UserSummary"
        role:
          $ref: "#/components/schemas/EnrollmentRole"
        status:
          $ref: "#/components/schemas/EnrollmentStatus"
        joinedAt:
          type:
            - integer
            - "null"
          format: int64
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64

    # ---------- Modules ----------
...
    ErrorResp:
      type: object
      properties:
        timestamp:
          type: integer
          format: int64
        path:
          type: string
        status:
          type: integer
        requestId:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
    ApiError:
      oneOf:
        - $ref: "#/components/schemas/InputError"
        - $ref: "#/components/schemas/GenericError"
      discriminator:
        propertyName: type
    InputError:
      type: object
      properties:
        type:
          type: string
          enum:
            - InputError
        field:
          type: string
        message:
          type: string
    GenericError:
      type: object
      properties:
        type:
          type: string
          enum:
            - SystemError
        reason:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation errors or malformed input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    Unauthorized:
      description: Unauthorized - missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
paths:
...
  /v1/courses:
    get:
      tags:
        - courses
      operationId: listCourses
      summary: List courses
      description: >
        Retrieves courses based on the specified type filter.
        Use type=created to list courses created by the user,
        type=enrolled for enrolled courses, or type=public for publicly visible courses.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - name: type
          in: query
          required: false
          description: Filter courses by relationship type
          schema:
            type: string
            enum:
              - created
              - enrolled
              - public
      responses:
        "200":
          description: List of courses matching the filter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags:
        - courses
      operationId: createCourse
      summary: Create a new course
      description: >
        Creates a new course. Requires the authenticated user to have the global
        UserRole.TEACHER role. The user becomes the course creator and is
        auto-enrolled as a TEACHER with ACTIVE status.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: User does not have the TEACHER global role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/courses/{id}:
    get:
      tags:
        - courses
      operationId: getCourse
      summary: Get a course by ID
      description: Retrieves the details of a specific course.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - name: id
          in: path
          required: true
          description: TypeID of the course
          schema:
            type: string
            format: typeid
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - courses
      operationId: updateCourse
      summary: Update a course
      description: Partially updates a course. Only the course creator or a teacher can update it.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - name: id
          in: path
          required: true
          description: TypeID of the course
          schema:
            type: string
            format: typeid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseRequest"
      responses:
        "200":
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - courses
      operationId: deleteCourse
      summary: Delete a course
      description: Deletes a course. Only the course creator can delete it.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - name: id
          in: path
          required: true
          description: TypeID of the course
          schema:
            type: string
            format: typeid
      responses:
        "204":
          description: Course deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/courses/code/{code}:
    get:
      tags:
        - courses
      operationId: getCourseByCode
      summary: Get a course by its unique code
      description: >
        Retrieves the details of a specific course using its unique 10-digit numeric code.
        Useful for sharing and joining courses via a short code instead of a full ID.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - name: code
          in: path
          required: true
          description: Unique 10-digit numeric code of the course
          schema:
            type: string
            pattern: "^\\d{10}$"
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ========================================================================
  # ENROLLMENTS
  # ========================================================================
  /v1/courses/{courseId}/enroll:
    post:
      tags:
        - enrollments
      operationId: enrollInCourse
      summary: Enroll in a course
      description: >
        Enrolls the authenticated user in the specified course.
        Optionally specify an enrollment role (defaults to LEARNER).
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/CourseIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentRequest"
      responses:
        "201":
          description: Enrollment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/courses/{courseId}/enrollments:
    get:
      tags:
        - enrollments
      operationId: listCourseEnrollments
      summary: List enrollments for a course
      description: >
        Retrieves all enrollments for a specific course.
        Typically restricted to teachers and the course creator.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/CourseIdPath"
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnrollmentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/courses/{courseId}/my-enrollment:
    get:
      tags:
        - enrollments
      operationId: getMyEnrollment
      summary: Get current user enrollment for a course
      description: Retrieves the enrollment record for the authenticated user in the specified course.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/CourseIdPath"
      responses:
        "200":
          description: Enrollment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/enrollments/{enrollmentId}:
    put:
      tags:
        - enrollments
      operationId: updateEnrollment
      summary: Update an enrollment
      description: Updates the role or status of an enrollment. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/EnrollmentIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnrollmentRequest"
      responses:
        "200":
          description: Enrollment updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - enrollments
      operationId: deleteEnrollment
      summary: Delete an enrollment
      description: Removes an enrollment. The user can unenroll themselves, or a teacher/creator can remove enrollments.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/EnrollmentIdPath"
      responses:
        "204":
          description: Enrollment deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/enrollments/{enrollmentId}/approve:
    post:
      tags:
        - enrollments
      operationId: approveEnrollment
      summary: Approve a pending enrollment
      description: Approves a pending enrollment request. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/EnrollmentIdPath"
      responses:
        "200":
          description: Enrollment approved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/enrollments/{enrollmentId}/reject:
    post:
      tags:
        - enrollments
      operationId: rejectEnrollment
      summary: Reject a pending enrollment
      description: Rejects a pending enrollment request. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/EnrollmentIdPath"
      responses:
        "200":
          description: Enrollment rejected successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ========================================================================
  # MODULES
  # ========================================================================
...